Add details to a map

Add details to a map

map_details(
  primeMeridian = "center",
  hires = FALSE,
  col = "darkolivegreen4",
  interior = FALSE,
  axes = FALSE,
  border = FALSE,
  boundaries.col = "black",
  grid = FALSE,
  grid.col = "white",
  grid.lwd = 0.5,
  cex.axis = 0.75,
  fill = TRUE,
  boundary = TRUE,
  water = NULL,
  land = TRUE,
  countries = FALSE,
  nx = NULL,
  ny = nx,
  labels = TRUE,
  ...
)

Arguments

primeMeridian

Prime Meridian used in the map, possible values are "center" and "left", the former is the default.

hires

Boolean, use high resolution coastline?

col

Color table to use for image, by default rev(rainbow(nlevel/10, start = 0/6, end = 4/6))

interior

If FALSE, interior segments are not drawn. An interior segment is a line segment of the map that bounds two of the polygons to be drawn. This argument is ignored if fill is TRUE.

axes

Boolean, add axes to the plot? Default is TRUE.

border

This only works if x and y are matrices – if NA the quadralaterals will have a border color that is the same as the interior color. Otherwise this specifies the color to use.

boundaries.col

Color of the country boundaries.

grid

Boolean, add grid lines over the map?

grid.col

Color of the grid.

grid.lwd

Width of the grid lines.

fill

logical flag that says whether to draw lines or fill areas. If FALSE, the lines bounding each region will be drawn (but only once, for interior lines). If TRUE, each region will be filled using colors from the col = argument, and bounding lines are drawn by default using par("fg"). To hide the bounding lines, use border=NA (see ...).

boundary

If FALSE, boundary segments are not drawn. A boundary segment is a line segment of the map that bounds only one of the polygons to be drawn. This argument is ignored if fill is TRUE.

water

Color of the water masses.

land

Boolean, add a polygon over the land? If FALSE, only coastline is added.

countries

Boolean, do add country boundaries? Default is FALSE.

nx, ny

number of cells of the grid in x and y direction. When NULL, as per default, the grid aligns with the tick marks on the corresponding default axis (i.e., tickmarks as computed by axTicks). When NA, no grid lines are drawn in the corresponding direction.

labels

Boolean, add LATITUDE and LONGITUDE labels?

...